Move Pearson hash function to cache package
